Fix clang-tidy and clang-format linting errors
- Add explicit static_cast<int> for narrowing conversions from
  sqlite3_int64 in db_manager.cpp (x2), int64_t in api_routes.cpp (x2),
  config_parser.cpp, and google_oauth.cpp
- Fix bugprone-misplaced-widening-cast in token_encryption.cpp:
  cast KEY_LEN to size_t before multiplication, and IV_LEN before addition
- Fix bugprone-implicit-widening-of-multiplication-result: cast loop index
  i to size_t before multiplying when indexing keyHex
- Fix narrowing conversion from unsigned char to char in two
  BASE64_CHARS.find() calls in token_encryption.cpp
- Fix middleware.h after_handle signature to use context& instead of
  const context& to match Crow's arity-3 type trait
- Apply clang-format to all touched files

+++ src/config_parser.cpp
@@ -34,7 +34,7 @@ Config loadConfig(const std::string &configFilePath) {
                       << ". Using defaults.\n";
         } else {
             if (jsonBody.has("port")) {
-                config.port = jsonBody["port"].i();
+                config.port = static_cast<int>(jsonBody["port"].i());
             } else {
                 std::cerr << "Notice: 'port' not found in config. Using default: " << config.port
                           << "\n";
